Extract structures from SaveGame.cpp
